[ fix ] Footer 위치 수정
